refactor: :chrone: rm few files
