+++ apps/kubegems/src/views/appstore/list.vue
@@ -214,7 +214,7 @@
         };
       default:
         return {
-          imgSrc: config.layout.LOGO_BLUE,
+          imgSrc: config.layout.LOGO_PRIMARY,
           tip: i18nLocal.t('tip.kubegems'),
         };
     }
