+++ apps/kubegems/src/views/resource/app/index.vue
@@ -204,7 +204,7 @@
                 height: '10px',
                 minWidth: '10px',
                 width: '10px',
-                backgroundColor: `${POD_STATUS_COLOR[item.status.phase] || config.theme.THEME_COLOR_OTHER.error}`,
+                backgroundColor: `${POD_STATUS_COLOR[item.status.phase] || config.theme.THEME_COLOR_EXTEND.error}`,
               }"
             />
             <span>
