+++ apps/kubegems/src/views/setting/model_registry/index.vue
@@ -186,7 +186,7 @@
         };
       default:
         return {
-          imgSrc: config.layout.LOGO_BLUE,
+          imgSrc: config.layout.LOGO_PRIMARY,
           tip: i18nLocal.t('tip.kubegems'),
           address: '',
         };
