+++ apps/servicemesh/src/config.json
@@ -2,7 +2,7 @@
   "layout": {
     "PLATFORM": "KubeGems",
     "MANUAL": "https://www.kubegems.io/docs/concepts/what-is-kubegems",
-    "LOGO_BLUE": "/logo-blue.svg",
+    "LOGO_PRIMARY": "/logo-blue.svg",
     "LOGO_WHITE": "/logo-white.svg",
     "SLOGAN_CN": "更便捷的云原生管理平台",
     "SLOGAN_EN": "Let cloudnative management more easily"
@@ -15,12 +15,13 @@
       "accent": "#fc4b6c",
       "default": "#563dea"
     },
-    "THEME_COLOR_OTHER": {
+    "THEME_COLOR_EXTEND": {
       "warning": "#fb8c00",
       "error": "#ff5252",
       "grey": "#607d8b",
       "greyLight": "#9e9e9e",
       "critical": "#673ab7"
-    }
+    },
+    "FONT_FAMILY": "zcool qingke huangyou, microsoft yahei, pingfang sc !default"
   }
 }
